requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# ── LINE Messaging API ────────────────────────────────────────────────────────
line-bot-sdk>=3.12.0

# ── NLP (Phase 1) ─────────────────────────────────────────────────────────────
pythainlp>=5.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
httpx>=0.27.0          # Required internally by line-bot-sdk v3

# ── Phase 2 (ML NER) — uncomment when ready ───────────────────────────────────
# transformers>=4.40.0
# torch>=2.2.0
# sentencepiece>=0.1.99
# accelerate>=0.27.0
